poprawki

+++ CzyDobre/Controllers/ProduktController.cs
@@ -13,7 +13,8 @@ namespace CzyDobre.Controllers
     {
         DBEntities db = new DBEntities();
 
-        // GET: Produkt
+        [Route("lista-produktow")]
+        [Route("Produkt/ProduktList")]
         [Authorize(Roles = "Admin, Moderator")]
         public ActionResult ProduktList()
         {
@@ -21,28 +22,60 @@ namespace CzyDobre.Controllers
             return View(produktList);
         }
 
+        [Route("edytowanie-produktu")]
+        [Route("Produkt/Edit")]
+        [Authorize(Roles = "Admin, Moderator")]
+        public ActionResult Edit(AspNetProduct obj)
+        {
+            if (obj != null)
+            {
+                return View(obj);
+            }
+            else
+            {
+                return View();
+            }
+            return View();
+        }
+
+        [Route("edytuj-produkt")]
+        [Route("Produkt/EditProdukt")]
+        [Authorize(Roles = "Admin, Moderator")]
+        public ActionResult EditProdukt(AspNetProduct obj)
+        {
+            return View();
+        }
 
         [Route("usuwanie-produktu")]
         [Route("Produkt/Delete")]
         [Authorize(Roles = "Admin")]
         public ActionResult Delete(int idProdukt, string nazwaProduktu)
         {
-            var idProduktu = db.AspNetProducts.Where(m => m.Id_Product == idProdukt).FirstOrDefault();
-            var idZdjecia = db.AspNetImages.Where(m => m.Id_Product == idProdukt).FirstOrDefault();
-
-            if (idProduktu != null)
+            try
             {
-                db.AspNetProducts.Remove(idProduktu);
-                if(idZdjecia != null)
+                var idProduktu = db.AspNetProducts.Where(m => m.Id_Product == idProdukt).FirstOrDefault();
+                var idZdjecia = db.AspNetImages.Where(m => m.Id_Product == idProdukt).FirstOrDefault();
+                if (idProduktu != null && idZdjecia != null)
                 {
+                    db.AspNetProducts.Remove(idProduktu);
                     db.AspNetImages.Remove(idZdjecia);
+                    db.SaveChanges();
+                    this.AddNotification("Produkt \"" + nazwaProduktu + "\" został pomyślnie usunięty!", NotificationType.SUCCESS);
                 }
-                db.SaveChanges();
-                this.AddNotification("Produkt \"" + nazwaProduktu + "\" został pomyślnie usunięty!", NotificationType.SUCCESS);
+                else
+                {
+                    this.AddNotification("Produkt nie został usunięty!", NotificationType.INFO);
+                }
+                var produktList = db.AspNetProducts.ToList();
+                return View("ProduktList", produktList);
             }
-            var produktList = db.AspNetProducts.ToList();
-            return View("ProduktList", produktList);
+            catch (Exception ex)
+            {
+                this.AddNotification($"Nie można usunać produktu, który posiada opinię", NotificationType.ERROR);
+                return RedirectToAction("ProduktList");
+            }
+            var produktListt = db.AspNetProducts.ToList();
+            return View("ProduktList", produktListt);
         }
-
     }
 }
